+++ src/server.ts
@@ -1,55 +1,7 @@
 import { ApolloServer } from '@apollo/server';
 import { startStandaloneServer } from '@apollo/server/standalone';
-
-interface User {
-  name: string;
-  email: string;
-  birthDate: string;
-}
-
-const typeDefs = `
-  type Query {
-    hello: String
-  }
-
-  type User {
-    id: ID!
-    name: String!
-    email: String!
-    birthDate: String!
-  }
-
-  input UserInput {
-    name: String!
-    email: String!
-    password: String!
-    birthDate: String!
-  }
-
-  type Mutation {
-    createUser(data: UserInput!): User
-  }
-`;
-
-const resolvers = {
-  Query: {
-    hello: (): string => {
-      return 'Hello world!';
-    },
-  },
-
-  Mutation: {
-    createUser: (parent, args: { data: User }) => {
-      const newUser = {
-        id: 1,
-        name: args.data.name,
-        email: args.data.email,
-        birthDate: args.data.birthDate,
-      };
-      return newUser;
-    },
-  },
-};
+import { typeDefs } from './type-defs';
+import { resolvers } from './resolvers';
 
 const server = new ApolloServer({
   typeDefs,
@@ -72,8 +24,6 @@ async function startServer() {
   }
 }
 
-startServer()
-  .then(() => {})
-  .catch((error) => {
-    console.error(`Error in execution: ${error}`);
-  });
+startServer().catch((error) => {
+  console.error(`Error in execution: ${error}`);
+});
